+++ NEWS.md
@@ -1,4 +1,9 @@
-# qualtRics 3.1.2.900 [WIP]
+# qualtRics (development version)
+
+- Add `fetch_distributions()` to access distribution data for a specific survey thanks to @dsen6644 (#169)
+- Handle mailing list embedded data better thanks to @dsen6644 (#175)
+
+# qualtRics 3.1.3
 
 - Update `include_questions` argument to use correct name in API request.
 - Build API payloads with jsonlite (#155) thanks to @jmobrien